refactor!: 以 OCCE(1.20.1 Forge) 源码树为基线整体重写，替换掉原 1.7.10 派生树
按用户要求：不再在 1.7.10 派生树上逐文件打补丁，而是**直接以 OCCE 的源码树为基线**
（mod_src/OpenComputers-CE-dev-MC1.20/src/main），搬到本项目后适配 NeoForge 1.21.1。
旧的派生树完整保存在提交 b94f915 / 11d776d 中，可随时对照。

== 本次完成的机械基线 ==

1. 源码整体替换：
   - scala 677 个（含 OCCE 混放在 src/main/scala 里的 35 个 .java）
   - java 192 个
   - 资源 2519 个（assets/opencomputers_neo + data/opencomputers_neo + src/data/resources 的 advancement）
2. 命名空间改造：mod id opencomputers → opencomputers_neo；
   `OpenComputers.ID`、`api/IMC.java` 的 MOD_ID、746+36 个资源 JSON 里的 `opencomputers:` → `opencomputers_neo:`，残留 0。
3. Forge → NeoForge 批量包名改造（904 个文件）：
   `net.minecraftforge.api.distmarker` → `net.neoforged.api.distmarker`（74 个文件）
   `net.minecraftforge.eventbus.api` → `net.neoforged.bus.api`（67）
   `net.minecraftforge.fml` → `net.neoforged.fml`（15）
   `MinecraftForge` → `NeoForge`（25）
   其余 `net.minecraftforge.*` → `net.neoforged.neoforge.*`（222）
4. 构建配置接到新树：
   - `gradle.properties` 的 `scala_ported_packages` 改为 `li/cil/oc/**`（整棵树都编译）
   - 新增 `src/main/templates/META-INF/neoforge.mods.toml`（NeoForge 格式，modLoader=javafml；
     OCCE 用的是 kotori_scala，那是 ScalaCatsForce 的加载器，我们 scala-library 已用 jarJar 打包）
   - 删除 OCCE 的 Forge 格式 `META-INF/mods.toml`
5. 1.21.1 物品 NBT 基础设施（NeoForge 已移除 ItemStack 上的 CompoundTag）：
   - `common/DataComponents.java`：注册自定义组件 `opencomputers_neo:nbt`，里面装**可变**的
     CompoundTag。刻意不用原版 `CUSTOM_DATA`（它取出的是副本，写回必须整份 set，会静默丢写入）
   - `util/ItemNBT.java`（Java 侧）+ `util/ExtendedItemStack.scala`（Scala 隐式，补回
     getTag/hasTag/setTag/getOrCreateTag）；37 个 Scala 文件已插入该隐式导入
   - `util/RegistryAccessHelper.java`：物品编解码所需的真实注册表访问器
     （绝不能用 RegistryAccess.EMPTY，否则每个物品都会被写成空标签、读档即丢失）
6. Java 层修到 **0 错误**：
   - `@Cancelable` → `ICancellableEvent`（9 个 event 文件）
   - `SideTracker`：`forgespi.Environment` → `EffectiveSide`（按逻辑侧判断，不能用物理侧）
   - `load`/`saveAdditional` → `loadAdditional`/`saveAdditional(CompoundTag, HolderLookup.Provider)`
   - `ItemStack.save(CompoundTag)` / `.of(tag)` / `.getOrCreateTag()` → 走上面那套新基础设施
7. 校验工具：`tools/scalac-check*.ps1` 现在会把 `src/main/scala` 下的 .java 也交给 scalac
   （联合编译：这些 Java 文件引用 Scala 类，不能进 compileJava，因为它先于 compileScala）

== 当前状态 ==
- `gradlew compileJava`：0 错误
- Scala：1192 个 parser/typer 错误（refchecks 尚未运行，真实积压更多）
  主要类别：RegistryObject → DeferredHolder、IForgeItem/IForgeBlock 已移除、
  capabilities 体系重写、ForgeRegistries → BuiltInRegistries、
  TickEvent 包路径、SimpleChannel → payload 体系、FMLJavaModLoadingContext、
  forgespi/Environment、neoforge.network.simple

== 下一步 ==
按错误类别从大到小推进：先做纯机械重命名（RegistryObject、ForgeRegistries、TickEvent 等），
再重写子系统（capabilities、网络、注册层、mod 入口）。

+++ src/main/scala/li/cil/oc/server/machine/luaj/OSAPI.scala
@@ -18,7 +18,7 @@ class OSAPI(owner: LuaJLuaArchitecture) extends LuaJAPI(owner) {
         else "%d/%m/%y %H:%M:%S"
       val time =
         if (args.narg > 1 && args.isnumber(2)) args.todouble(2)
-        else (machine.worldTime + 6000) * 60 * 60 / 1000
+        else ((machine.worldTime + 6000) * 60 * 60 / 1000).toDouble
 
       val dt = GameTimeFormatter.parse(time)
       def fmt(format: String) = {
